# internal/app/core/transport/rest/users-router/users-router.go
package users_router

import (
	"errors"
	users_entities "mkk_basis/rest_api/internal/app/core/entities/users-entities"
	users_filter "mkk_basis/rest_api/internal/app/core/entities/users-filter"
	users_handler "mkk_basis/rest_api/internal/app/core/handlers/rest/users-handler"
	"mkk_basis/rest_api/internal/common"
	auth_middleware "mkk_basis/rest_api/internal/components/http/middlewares"
	"net/http"
	"strconv"
	"time"

	"github.com/gin-gonic/gin"
)

func AddRoutes(r *gin.RouterGroup) {
	router := r.Group("/users")
	router.Use(auth_middleware.AuthMiddleware())
	{
		router.POST("", createUserRoute)
		router.GET("", getUsersRoute)
		router.GET("/filter", getUsersFilterRoute)
		router.GET("/by-username", getUserByUsernameRoute)
		router.GET("/top-task-creators", getTopTaskCreatorsRoute)

		router.GET("/:id", getUserRoute)
		router.PUT("/:id", updateUserRoute)
		router.DELETE("/:id", deleteUserRoute)
	}
}

// Create user
//
//	@Summary		Create user
//	@Description	Создание пользователя
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			user	body		users_entities.UserRequest	true	"User"
//	@Success		201		{object}	common.APIResponse{result=users_entities.UserResponse}
//	@Router			/api/v1/users [post]
func createUserRoute(c *gin.Context) {
	var request *users_entities.UserRequest

	if err := c.ShouldBindJSON(&request); err != nil {
		c.JSON(http.StatusBadRequest, common.ErrorResponse(err))
		return
	}

	user, err := users_handler.CreateUser(c.Request.Context(), request)
	if err != nil {
		c.JSON(http.StatusBadRequest, common.ErrorResponse(err))
		return
	}

	c.JSON(http.StatusCreated, common.ResultResponseNoErr(user))
}

// Update user
//
//	@Summary		Update user
//	@Description	Обновление пользователя
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			id		path		int							true	"User ID"
//	@Param			user	body		users_entities.UserRequest	true	"User"
//	@Success		200		{object}	common.APIResponse{result=users_entities.UserResponse}
//	@Router			/api/v1/users/{id} [put]
func updateUserRoute(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseUint(idStr, 10, 64)
	if err != nil {
		c.JSON(http.StatusBadRequest, common.ErrorResponse(err))
		return
	}

	var request *users_entities.UserRequest

	if err = c.ShouldBindJSON(&request); err != nil {
		c.JSON(http.StatusBadRequest, common.ErrorResponse(err))
		return
	}

	user, err := users_handler.UpdateUser(c.Request.Context(), id, request)
	if err != nil {
		c.JSON(http.StatusBadRequest, common.ErrorResponse(err))
		return
	}

	c.JSON(http.StatusOK, common.ResultResponseNoErr(user))
}

// Get users
//
//	@Summary		Get users
//	@Description	Получить список пользователей без фильтра
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Success		200	{object}	common.APIResponse{result=[]users_entities.UserResponse}
//	@Router			/api/v1/users [get]
func getUsersRoute(c *gin.Context) {
	users, err := users_handler.GetUsers(c.Request.Context())
	if err != nil {
		c.JSON(http.StatusBadRequest, common.ErrorResponse(err))
		return
	}

	c.JSON(http.StatusOK, common.ResultResponseNoErr(users))
}

// Get users with filter
//
//	@Summary		Get users with filter
//	@Description	Получить список пользователей с фильтрацией и пагинацией
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			query	query		users_filter.UserFilterRequest	true	"Filter parameters"
//	@Success		200		{object}	common.APIResponse{result=[]users_entities.UserResponse}
//	@Router			/api/v1/users/filter [get]
func getUsersFilterRoute(c *gin.Context) {
	var params users_filter.UserFilterRequest

	if err := c.ShouldBindQuery(&params); err != nil {
		c.JSON(http.StatusBadRequest, common.ErrorResponse(err))
		return
	}

	usersResp, err := users_handler.GetUsersFilter(c.Request.Context(), &params)
	if err != nil {
		c.JSON(http.StatusBadRequest, common.ErrorResponse(err))
		return
	}

	c.Header("Content-Range", strconv.FormatInt(usersResp.ContentRange, 10))
	c.JSON(http.StatusOK, common.ResultResponseNoErr(usersResp.Users))
}

// Get top task creators
//
//	@Summary		Get top task creators by team
//	@Description	Получить топ-3 пользователей по количеству созданных задач в каждой команде за указанный месяц
//	@Tags			users
//	@Produce		json
//	@Param			month	query		string	true	"Month in YYYY-MM format"
//	@Success		200		{object}	common.APIResponse{result=[]users_entities.TeamTopTaskCreatorResponse}
//	@Router			/api/v1/users/top-task-creators [get]
func getTopTaskCreatorsRoute(c *gin.Context) {
	month, err := time.Parse("2006-01", c.Query("month"))
	if err != nil {
		c.JSON(http.StatusBadRequest, common.ErrorResponse(errors.New("month must be in YYYY-MM format")))
		return
	}

	creators, err := users_handler.GetTopTaskCreatorsByTeamForMonth(c.Request.Context(), month)
	if err != nil {
		c.JSON(http.StatusInternalServerError, common.ErrorResponse(err))
		return
	}

	c.JSON(http.StatusOK, common.ResultResponseNoErr(creators))
}

// Get user by id
//
//	@Summary		Get user by id
//	@Description	Получить пользователя по ID
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			id	path		int	true	"User ID"
//	@Success		200	{object}	common.APIResponse{result=users_entities.UserResponse}
//	@Router			/api/v1/users/{id} [get]
func getUserRoute(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseUint(idStr, 10, 64)
	if err != nil {
		c.JSON(http.StatusBadRequest, common.ErrorResponse(err))
		return
	}

	user, err := users_handler.GetUserById(c.Request.Context(), id)
	if err != nil {
		c.JSON(http.StatusBadRequest, common.ErrorResponse(err))
		return
	}
	c.JSON(http.StatusOK, common.ResultResponseNoErr(user))
}

// Get user by username
//
//	@Summary		Get user by username
//	@Description	Получить пользователя по username
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			username	query		string	true	"User username"
//	@Success		200			{object}	common.APIResponse{result=users_entities.UserResponse}
//	@Router			/api/v1/users/by-username [get]
func getUserByUsernameRoute(c *gin.Context) {
	username := c.Query("username")
	if username == "" {
		err := errors.New("username is required")
		c.JSON(http.StatusBadRequest, common.ErrorResponse(err))
		return
	}

	user, err := users_handler.GetUserByUsername(c.Request.Context(), username)
	if err != nil {
		c.JSON(http.StatusBadRequest, common.ErrorResponse(err))
		return
	}
	c.JSON(http.StatusOK, common.ResultResponseNoErr(user))
}

// Delete user by id
//
//	@Summary		Delete user by id
//	@Description	Удалить пользователя по ID
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			id	path		int	true	"User ID"
//	@Success		200	{object}	common.APIResponse{result=uint64}
//	@Router			/api/v1/users/{id} [delete]
func deleteUserRoute(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.ParseUint(idStr, 10, 64)
	if err != nil {
		c.JSON(http.StatusBadRequest, common.ErrorResponse(err))
		return
	}

	returnedUserID, err := users_handler.DeleteUser(c.Request.Context(), id)
	if err != nil {
		c.JSON(http.StatusBadRequest, common.ErrorResponse(err))
		return
	}

	c.JSON(http.StatusOK, common.ResultResponseNoErr(returnedUserID))
}